# documentation/api/openapi.json
{
  "basePath": "/api/v1",
  "consumes": [
    "application/json"
  ],
  "definitions": {
...
    "CharacterCreate": {
      "properties": {
        "avatar_image": {
          "description": "Character avatar image URL (for JSON requests)",
          "type": "string"
        },
        "description": {
          "description": "Character description",
          "type": "string"
        },
        "label": {
          "description": "Unique character identifier",
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "name": {
          "description": "Character name",
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        }
      },
      "required": [
        "label",
        "name"
      ],
      "type": "object"
    },
    "CharacterUpdate": {
      "properties": {
        "avatar_image": {
          "description": "Character avatar image path",
          "type": "string"
        },
        "description": {
          "description": "Character description",
          "type": "string"
        },
        "label": {
          "description": "Unique character identifier",
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "name": {
          "description": "Character name",
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "Response": {
      "properties": {
        "data": {
          "description": "Response data",
          "type": "object"
        },
        "error": {
          "description": "Error information, if any",
          "type": "object"
        },
        "meta": {
          "description": "Additional metadata",
          "type": "object"
        },
        "success": {
          "default": true,
          "description": "Success status",
          "type": "boolean"
        }
      },
      "type": "object"
    },
...
  },
  "info": {
    "description": "API for the Roleplay Chat Web App",
    "title": "Roleplay Chat API",
    "version": "1.0"
  },
  "paths": {
...
    "/characters/": {
      "get": {
        "operationId": "list_characters",
        "parameters": [
          {
            "default": 1,
            "description": "Page number",
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "type": "integer"
          },
          {
            "description": "An optional fields mask",
            "format": "mask",
            "in": "header",
            "name": "X-Fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          }
        },
        "summary": "List all characters with pagination",
        "tags": [
          "characters"
        ]
      },
      "post": {
        "description": "## Request Formats:\n\n### 1. JSON (Content-Type: application/json)\nStandard character creation with avatar_image as URL string.\nUses the CharacterCreate schema.\n\n### 2. Multipart Form (Content-Type: multipart/form-data)  \nCharacter creation with file upload support:\n- **label**: string (required) - Unique character identifier\n- **name**: string (required) - Character display name\n- **description**: string (optional) - Character description\n- **avatar_image**: file (optional) - Image file (PNG, JPG, GIF, WebP, max 5MB, max 1024x1024px)\n\n## File Upload Details:\n- Supported formats: PNG, JPG, JPEG, GIF, WebP\n- Maximum file size: 5MB\n- Maximum dimensions: 1024x1024 pixels\n- Large images are automatically resized\n- Files are stored securely with UUID-generated names",
        "operationId": "create_character",
        "parameters": [
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CharacterCreate"
            }
          },
          {
            "default": "application/json",
            "description": "Content type of the request",
            "enum": [
              "application/json",
              "multipart/form-data"
            ],
            "in": "header",
            "name": "Content-Type",
            "type": "string"
          },
          {
            "description": "An optional fields mask",
            "format": "mask",
            "in": "header",
            "name": "X-Fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "201": {
            "description": "Character created successfully"
          },
          "400": {
            "description": "Bad request - validation error or invalid file"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create a new character with optional avatar image upload",
        "tags": [
          "characters"
        ]
      }
    },
    "/characters/search": {
      "get": {
        "operationId": "search_characters",
        "parameters": [
          {
            "description": "Search query",
            "in": "query",
            "name": "query",
            "required": true,
            "type": "string"
          },
          {
            "default": 1,
            "description": "Page number",
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "type": "integer"
          },
          {
            "description": "An optional fields mask",
            "format": "mask",
            "in": "header",
            "name": "X-Fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          }
        },
        "summary": "Search for characters by name or description",
        "tags": [
          "characters"
        ]
      }
    },
    "/characters/{id}": {
      "delete": {
        "operationId": "delete_character",
        "parameters": [
          {
            "description": "An optional fields mask",
            "format": "mask",
            "in": "header",
            "name": "X-Fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "summary": "Delete a character",
        "tags": [
          "characters"
        ]
      },
      "get": {
        "operationId": "get_character",
        "parameters": [
          {
            "description": "An optional fields mask",
            "format": "mask",
            "in": "header",
            "name": "X-Fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "summary": "Get a character by ID",
        "tags": [
          "characters"
        ]
      },
      "parameters": [
        {
          "description": "The character identifier",
          "in": "path",
          "name": "id",
          "required": true,
          "type": "integer"
        }
      ],
      "put": {
        "operationId": "update_character",
        "parameters": [
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CharacterUpdate"
            }
          },
          {
            "description": "An optional fields mask",
            "format": "mask",
            "in": "header",
            "name": "X-Fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "404": {
            "description": "Character not found"
          }
        },
        "summary": "Update a character",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/chat-sessions/character/{character_id}": {
      "get": {
        "operationId": "get_chat_sessions_by_character",
        "parameters": [
          {
            "description": "An optional fields mask",
            "format": "mask",
            "in": "header",
            "name": "X-Fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          }
        },
        "summary": "Get chat sessions for a specific character",
        "tags": [
          "chat-sessions"
        ]
      },
      "parameters": [
        {
          "description": "The character identifier",
          "in": "path",
          "name": "character_id",
          "required": true,
          "type": "integer"
        }
      ]
    },
...
  },
  "produces": [
    "application/json"
  ],
...
  "swagger": "2.0",
...
}